+++ programs/mpl-core/src/error.rs
@@ -177,27 +177,23 @@ pub enum MplCoreError {
     #[error("Invalid plugin operation")]
     InvalidPluginOperation,
 
-    /// 42 - Collection must be empty to be burned
-    #[error("Collection must be empty to be burned")]
-    CollectionMustBeEmpty,
-
-    /// 43 - Two data sources provided, only one is allowed
+    /// 42 - Two data sources provided, only one is allowed
     #[error("Two data sources provided, only one is allowed")]
     TwoDataSources,
 
-    /// 44 - External Plugin does not support this operation
+    /// 43 - External Plugin does not support this operation
     #[error("External Plugin does not support this operation")]
     UnsupportedOperation,
 
-    /// 45 - No data sources provided, one is required
+    /// 44 - No data sources provided, one is required
     #[error("No data sources provided, one is required")]
     NoDataSources,
 
-    /// 46 - This plugin adapter cannot be added to an Asset
+    /// 45 - This plugin adapter cannot be added to an Asset
     #[error("This plugin adapter cannot be added to an Asset")]
     InvalidPluginAdapterTarget,
 
-    /// 47 - Cannot add a Data Section without a linked external plugin
+    /// 46 - Cannot add a Data Section without a linked external plugin
     #[error("Cannot add a Data Section without a linked external plugin")]
     CannotAddDataSection,
 }
